% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{create_custom_action_type}
\alias{create_custom_action_type}
\title{Creates a new custom action that can be used in all pipelines associated with the AWS account}
\usage{
create_custom_action_type(category, provider, version, inputArtifactDetails,
  outputArtifactDetails, settings = NULL, configurationProperties = NULL)
}
\arguments{
\item{category}{The category of the custom action, such as a build action or a test action.

Although Source and Approval are listed as valid values, they are not currently functional. These values are reserved for future use.}

\item{provider}{The provider of the service used in the custom action, such as AWS CodeDeploy.}

\item{version}{The version identifier of the custom action.}

\item{inputArtifactDetails}{The details of the input artifact for the action, such as its commit ID.}

\item{outputArtifactDetails}{The details of the output artifact of the action, such as its commit ID.}

\item{settings}{Returns information about the settings for an action type.}

\item{configurationProperties}{The configuration properties for the custom action.

You can refer to a name in the configuration properties of the custom action within the URL templates by following the format of {Config:name}, as long as the configuration property is both required and not secret. For more information, see \href{http://docs.aws.amazon.com/codepipeline/latest/userguide/how-to-create-custom-action.html}{Create a Custom Action for a Pipeline}.}
}
\description{
Creates a new custom action that can be used in all pipelines associated with the AWS account. Only used for custom actions.
}
\examples{

}
