% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{create_pipeline}
\alias{create_pipeline}
\title{Creates a pipeline}
\usage{
create_pipeline(pipeline)
}
\arguments{
\item{pipeline}{Represents the structure of actions and stages to be performed in the pipeline.}
}
\description{
Creates a pipeline.
}
\examples{

}
