% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{delete_custom_action_type}
\alias{delete_custom_action_type}
\title{Marks a custom action as deleted}
\usage{
delete_custom_action_type(
  category = "Source"|"Build"|"Deploy"|"Test"|"Invoke"|"Approval",
  provider = "string",
  version = "string"
)
}
\arguments{
\item{category}{[required] The category of the custom action that you want to delete, such as source or deploy.}

\item{provider}{[required] The provider of the service used in the custom action, such as AWS CodeDeploy.}

\item{version}{[required] The version of the custom action to delete.}
}
\description{
Marks a custom action as deleted. PollForJobs for the custom action will fail after the action is marked for deletion. Only used for custom actions.
}
\details{
To re-create a custom action after it has been deleted you must use a string in the version field that has never been used before. This string can be an incremented version number, for example. To restore a deleted custom action, use a JSON file that is identical to the deleted action, including the original string in the version field.
}
\examples{

}
