% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{delete_pipeline}
\alias{delete_pipeline}
\title{Deletes the specified pipeline}
\usage{
delete_pipeline(name)
}
\arguments{
\item{name}{The name of the pipeline to be deleted.}
}
\description{
Deletes the specified pipeline.
}
\examples{

}
