% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{disable_stage_transition}
\alias{disable_stage_transition}
\title{Prevents artifacts in a pipeline from transitioning to the next stage in the pipeline}
\usage{
disable_stage_transition(
  pipelineName = "string",
  stageName = "string",
  transitionType = "Inbound"|"Outbound",
  reason = "string"
)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline in which you want to disable the flow of artifacts from one stage to another.}

\item{stageName}{[required] The name of the stage where you want to disable the inbound or outbound transition of artifacts.}

\item{transitionType}{[required] Specifies whether artifacts will be prevented from transitioning into the stage and being processed by the actions in that stage (inbound), or prevented from transitioning from the stage after they have been processed by the actions in that stage (outbound).}

\item{reason}{[required] The reason given to the user why a stage is disabled, such as waiting for manual approval or manual tests. This message is displayed in the pipeline console UI.}
}
\description{
Prevents artifacts in a pipeline from transitioning to the next stage in the pipeline.
}
\examples{

}
