% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Returns the metadata, structure, stages, and actions of a pipeline}
\usage{
get_pipeline(name, version = NULL)
}
\arguments{
\item{name}{[required] The name of the pipeline for which you want to get information. Pipeline names must be unique under an Amazon Web Services (AWS) user account.}

\item{version}{The version number of the pipeline. If you do not specify a version, defaults to the most current version.}
}
\description{
Returns the metadata, structure, stages, and actions of a pipeline. Can be used to return the entire structure of a pipeline in JSON format, which can then be modified and used to update the pipeline structure with UpdatePipeline.
}
\section{Accepted Parameters}{
\preformatted{get_pipeline(
  name = "string",
  version = 123
)
}
}

