% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{get_pipeline_state}
\alias{get_pipeline_state}
\title{Returns information about the state of a pipeline, including the stages and actions}
\usage{
get_pipeline_state(
  name = "string"
)
}
\arguments{
\item{name}{[required] The name of the pipeline about which you want to get information.}
}
\description{
Returns information about the state of a pipeline, including the stages and actions.
}
\examples{

}
