% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{get_third_party_job_details}
\alias{get_third_party_job_details}
\title{Requests the details of a job for a third party action}
\usage{
get_third_party_job_details(jobId, clientToken)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID used for identifying the job.}

\item{clientToken}{[required] The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.}
}
\description{
Requests the details of a job for a third party action. Only used for partner actions.
}
\details{
When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.
}
\section{Accepted Parameters}{
\preformatted{get_third_party_job_details(
  jobId = "string",
  clientToken = "string"
)
}
}

