% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{list_action_types}
\alias{list_action_types}
\title{Gets a summary of all AWS CodePipeline action types associated with your account}
\usage{
list_action_types(
  actionOwnerFilter = "AWS"|"ThirdParty"|"Custom",
  nextToken = "string"
)
}
\arguments{
\item{actionOwnerFilter}{Filters the list of action types to those created by a specified entity.}

\item{nextToken}{An identifier that was returned from the previous list action types call, which can be used to return the next set of action types in the list.}
}
\description{
Gets a summary of all AWS CodePipeline action types associated with your account.
}
\examples{

}
