% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{list_pipeline_executions}
\alias{list_pipeline_executions}
\title{Gets a summary of the most recent executions for a pipeline}
\usage{
list_pipeline_executions(pipelineName, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{pipelineName}{The name of the pipeline for which you want to get execution summary information.}

\item{maxResults}{The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. The available pipeline execution history is limited to the most recent 12 months, based on pipeline execution start times. Default value is 100.}

\item{nextToken}{The token that was returned from the previous ListPipelineExecutions call, which can be used to return the next set of pipeline executions in the list.}
}
\description{
Gets a summary of the most recent executions for a pipeline.
}
\examples{

}
