% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_package.R
\docType{package}
\name{paws.codepipeline-package}
\alias{paws.codepipeline}
\alias{paws.codepipeline-package}
\title{paws.codepipeline: AWS CodePipeline}
\description{
This is the AWS CodePipeline API Reference. This
    guide provides descriptions of the actions and data types for AWS
    CodePipeline. Some functionality for your pipeline is only
    configurable through the API. For additional information, see the [AWS
    CodePipeline User
    Guide](http://docs.aws.amazon.com/codepipeline/latest/userguide/welcome.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
