% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{poll_for_third_party_jobs}
\alias{poll_for_third_party_jobs}
\title{Determines whether there are any third party jobs for a job worker to act on}
\usage{
poll_for_third_party_jobs(
  actionTypeId = list(
    category = "Source"|"Build"|"Deploy"|"Test"|"Invoke"|"Approval",
    owner = "AWS"|"ThirdParty"|"Custom",
    provider = "string",
    version = "string"
  ),
  maxBatchSize = 123
)
}
\arguments{
\item{actionTypeId}{[required] Represents information about an action type.}

\item{maxBatchSize}{The maximum number of jobs to return in a poll for jobs call.}
}
\description{
Determines whether there are any third party jobs for a job worker to act on. Only used for partner actions.
}
\details{
When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts.
}
\examples{

}
