% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{put_action_revision}
\alias{put_action_revision}
\title{Provides information to AWS CodePipeline about new revisions to a source}
\usage{
put_action_revision(pipelineName, stageName, actionName, actionRevision)
}
\arguments{
\item{pipelineName}{The name of the pipeline that will start processing the revision to the source.}

\item{stageName}{The name of the stage that contains the action that will act upon the revision.}

\item{actionName}{The name of the action that will process the revision.}

\item{actionRevision}{Represents information about the version (or revision) of an action.}
}
\description{
Provides information to AWS CodePipeline about new revisions to a source.
}
\examples{

}
