% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{put_approval_result}
\alias{put_approval_result}
\title{Provides the response to a manual approval request to AWS CodePipeline}
\usage{
put_approval_result(pipelineName, stageName, actionName, result, token)
}
\arguments{
\item{pipelineName}{The name of the pipeline that contains the action.}

\item{stageName}{The name of the stage that contains the action.}

\item{actionName}{The name of the action for which approval is requested.}

\item{result}{Represents information about the result of the approval request.}

\item{token}{The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState action and is used to validate that the approval request corresponding to this token is still valid.}
}
\description{
Provides the response to a manual approval request to AWS CodePipeline. Valid responses include Approved and Rejected.
}
\examples{

}
