% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{put_third_party_job_failure_result}
\alias{put_third_party_job_failure_result}
\title{Represents the failure of a third party job as returned to the pipeline by a job worker}
\usage{
put_third_party_job_failure_result(jobId, clientToken, failureDetails)
}
\arguments{
\item{jobId}{The ID of the job that failed. This is the same ID returned from PollForThirdPartyJobs.}

\item{clientToken}{The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.}

\item{failureDetails}{Represents information about failure details.}
}
\description{
Represents the failure of a third party job as returned to the pipeline by a job worker. Only used for partner actions.
}
\examples{

}
