% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{register_webhook_with_third_party}
\alias{register_webhook_with_third_party}
\title{Configures a connection between the webhook that was created and the external tool with events to be detected}
\usage{
register_webhook_with_third_party(webhookName = NULL)
}
\arguments{
\item{webhookName}{The name of an existing webhook created with PutWebhook to register with a supported third party.}
}
\description{
Configures a connection between the webhook that was created and the external tool with events to be detected.
}
\examples{

}
