% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{retry_stage_execution}
\alias{retry_stage_execution}
\title{Resumes the pipeline execution by retrying the last failed actions in a stage}
\usage{
retry_stage_execution(pipelineName, stageName, pipelineExecutionId, retryMode)
}
\arguments{
\item{pipelineName}{The name of the pipeline that contains the failed stage.}

\item{stageName}{The name of the failed stage to be retried.}

\item{pipelineExecutionId}{The ID of the pipeline execution in the failed stage to be retried. Use the GetPipelineState action to retrieve the current pipelineExecutionId of the failed stage}

\item{retryMode}{The scope of the retry attempt. Currently, the only supported value is FAILED\_ACTIONS.}
}
\description{
Resumes the pipeline execution by retrying the last failed actions in a stage.
}
\examples{

}
