% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{update_pipeline}
\alias{update_pipeline}
\title{Updates a specified pipeline with edits or changes to its structure}
\usage{
update_pipeline(pipeline)
}
\arguments{
\item{pipeline}{The name of the pipeline to be updated.}
}
\description{
Updates a specified pipeline with edits or changes to its structure. Use a JSON file with the pipeline structure in conjunction with UpdatePipeline to provide the full structure of the pipeline. Updating the pipeline increases the version number of the pipeline by 1.
}
\examples{

}
