% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{put_job_failure_result}
\alias{put_job_failure_result}
\title{Represents the failure of a job as returned to the pipeline by a job worker}
\usage{
put_job_failure_result(jobId, failureDetails)
}
\arguments{
\item{jobId}{The unique system-generated ID of the job that failed. This is the same ID returned from PollForJobs.}

\item{failureDetails}{The details about the failure of a job.}
}
\description{
Represents the failure of a job as returned to the pipeline by a job worker. Only used for custom actions.
}
\examples{

}
