% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codepipeline_operations.R
\name{start_pipeline_execution}
\alias{start_pipeline_execution}
\title{Starts the specified pipeline}
\usage{
start_pipeline_execution(name, clientRequestToken = NULL)
}
\arguments{
\item{name}{[required] The name of the pipeline to start.}

\item{clientRequestToken}{The system-generated unique ID used to identify a unique execution request.}
}
\description{
Starts the specified pipeline. Specifically, it begins processing the latest commit to the source location specified as part of the pipeline.
}
\section{Accepted Parameters}{
\preformatted{start_pipeline_execution(
  name = "string",
  clientRequestToken = "string"
)
}
}

